+++ src/object/object.cpp
@@ -19,7 +19,7 @@ static Obj* allocateObject(size_t size, ObjType type) {
     return object;
 }
 
-static ObjString* allocateString(char* chars, int length, uint32_t hash) {
+static ObjString* allocateString(char* chars, size_t length, size_t hash) {
     ObjString* string = ALLOCATE_OBJ(ObjString, OBJ_STRING);
     string->length = length;
     string->chars = chars;
@@ -28,7 +28,7 @@ static ObjString* allocateString(char* chars, int length, uint32_t hash) {
     return string;
 }
 
-static uint32_t hashString(const char* key, int length) {
+static uint32_t hashString(const char* key, size_t length) {
     uint32_t hash = 2166136261u;
     for (int i = 0; i < length; i++) {
         hash ^= key[i];
@@ -38,7 +38,7 @@ static uint32_t hashString(const char* key, int length) {
 }
 
 ObjString* copyString(const char* chars, size_t length) {
-    uint32_t hash = hashString(chars, length);
+    size_t hash = hashString(chars, length);
 
     ObjString* interned = HashTable::findString(&vm.strings, chars, length, hash);
     if (interned != nullptr) return interned;
